correction part 1

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: mbazirea <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/31 11:23:33 by mbazirea          #+#    #+#             */
-/*   Updated: 2022/10/31 11:46:54 by mbazirea         ###   ########.fr       */
+/*   Updated: 2022/11/02 13:28:28 by mbazirea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
